+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Prowide Core - CHANGELOG
 
-#### 10.33 - SNAPSHOT
+#### 10.3.3 - SNAPSHOT
+  * (PW-2541) Updated the narrative resolver, format field 70 (used in field 70), to allow multiple lines and using // as codeword separator (/CODEWORD1/text///CODEWORD2/text)
   * Adding a new entry to SchemeConstantsR class
+  * Fix: Enhanced the `DefaultMtMetadataStrategy` to prevent NPE when the message headers are malformed
 
 #### 10.3.2 - June 2025
   * Fix: Updated components for Field 13W